+++ modules/.prettierignore
@@ -44,7 +44,8 @@
 ##
 
     #
-    # This is a generated third-party file, not intended for editing or
+    # These are generated third-party files, not intended for editing or
     # inspection.
     #
+    /apps/frontend-js/frontend-js-svg4everybody-web/src/main/resources/META-INF/resources/svg4everybody.js
     /yarn-*.js
